move FacetFWsGeneratorTask and SurfCalcOptimizer to appropriate mdoules
