+++ atomate/vasp/workflows/base/surface.py
@@ -22,21 +22,11 @@ __version__ = "0.2"
 __email__ = "rit001@eng.ucsd.edu"
 __date__ = "11/30/17"
 
-import json
 import os
 
-from atomate.vasp.firetasks.run_calc import RunVaspCustodian
-from atomate.common.firetasks.glue_tasks import CreateFolder, RenameFile
-from atomate.vasp.firetasks.write_inputs import WriteVaspFromIOSet
-from atomate.vasp.firetasks.parse_outputs import VaspToDb
+from pymatgen import Structure
 
-from pymatgen.core.surface import generate_all_slabs, Structure, SlabGenerator, \
-    ReconstructionGenerator
-from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
-from pymatgen.io.vasp.sets import MVLSlabSet
-
-from fireworks.core.firework import Firework, Workflow, FiretaskBase, FWAction
-from fireworks import explicit_serialize
+from fireworks import Workflow
 
 
 class SurfaceWorkflowManager(object):
@@ -172,324 +162,3 @@ class SurfaceWorkflowManager(object):
                                            vsize=vsize, mpid=mpid, **kwargs)])
 
 
-class SurfCalcOptimizer(Firework):
-    """
-    Optimizer firework for surface property calculation and insertion. Will build
-        a Firework tailored to the type of structure and as a result, use  different
-        input sets and insert different additionals depending on the structure_type.
-    """
-
-    def __init__(self, structure, scratch_dir, k_product, db_file,
-                 vasp_cmd, structure_type, miller_index=[], scale_factor=[],
-                 vsize=None, mmi=None, ouc=None, shift=None, ssize=None,
-                 reconstruction=None, cwd=os.getcwd(), mpid="--", **kwargs):
-        """
-        Initializes the Firework.
-
-        Args:
-            structure (Structure): Can be a conventional unit cell, oriented
-                unit cell or slab cell.
-            scratch_dir (str): - if specified, uses this directory as the root
-                scratch dir. Supports env_chk.
-            k_product (int): Default to 50, kpoint number * length for a & b
-                directions, also for c direction in bulk calculations.
-            db_file (str): FULL path to file containing the database credentials.
-                Supports env_chk.
-            vasp_cmd (str): Command to run vasp.
-            structure_type (str): Type of structure. Options are:
-                conventional_unit_cell, oriented_unit_cell and slab_cell
-            miller_index ([h, k, l]): Miller index of plane parallel to
-                surface (and oriented unit cell).
-            ouc (Structure): The oriented_unit_cell from which
-                this Slab is created (by scaling in the c-direction).
-            shift (float): The shift in the c-direction applied to get the
-                termination.
-            ssize (float): Minimum slab size in Angstroms or number of hkl planes
-            vsize (float): Minimum vacuum size in Angstroms or number of hkl planes
-            mmi (int): Max Miller index.
-            scale_factor (array): Final computed scale factor that brings
-                the parent cell to the surface cell.
-            mpid (str): Materials Project ID of the conventional unit cell.
-            reconstruction (str): The name of the reconstruction
-                (if it is a reconstructed slab).
-            cwd (str): Location of directory to operate the
-                workflow and store the final outputs
-            \*\*kwargs: Other kwargs that are passed to Firework.__init__.
-        """
-
-        self.structure = structure
-        self.el = self.structure[0].species_string
-        self.structure_type = structure_type
-        self.k_product = k_product
-        self.mpid = mpid
-        self.hkl = miller_index
-        self.reconstruction = reconstruction
-        self.ssize = ssize
-        self.vsize = vsize
-        self.shift = shift
-        self.sg = SpacegroupAnalyzer(ouc if structure_type ==
-                                            "slab_cell" else structure)
-        self.scale_factor = scale_factor
-        self.ouc = ouc
-        self.cwd = cwd
-        self.vasp_cmd = vasp_cmd
-        self.scratch_dir = scratch_dir
-        self.db_file = db_file
-        self.mmi = mmi
-
-        super(SurfCalcOptimizer, self).__init__(self.get_tasks,
-                                                name=self.get_name, **kwargs)
-
-    @property
-    def get_input_set(self):
-
-        # Get the MVLSlabSet based on structure type
-        # Returns input_set
-
-        if self.structure_type != "slab_cell":
-            return MVLSlabSet(self.structure, bulk=False,
-                              k_product=self.k_product, get_locpot=True)
-        else:
-            return MVLSlabSet(self.structure, bulk=True,
-                              k_product=self.k_product)
-
-    @property
-    def get_name(self):
-
-        # Get the name of the calculation based on element,
-        # mpid, kproduct, structure type, and slab parameters
-        # Returns str (name)
-
-        if self.structure_type == "conventional_unit_cell":
-            return "%s_%s_conventional_unit_cell_k%s" % \
-                   (self.el, self.mpid, self.k_product)
-        elif self.structure_type == "oriented_unit_cell":
-            if self.reconstruction:
-                return "%s_%s_bulk_rec_k%s_%s" % \
-                       (self.el, self.mpid, self.k_product,
-                        self.reconstruction)
-            else:
-                return "%s_%s_bulk_k%s_%s%s%s" % \
-                       (self.el, self.mpid, self.k_product,
-                        self.hkl[0], self.hkl[1], self.hkl[2])
-        elif self.structure_type == "slab_cell":
-            if not self.reconstruction:
-                return "%s_%s_slab_k%s_s%sv%s_%s%s%s_shift%s" % \
-                       (self.el, self.mpid, self.k_product, self.ssize, self.vsize,
-                        self.hkl[0], self.hkl[1], self.hkl[2], self.shift)
-            else:
-                return "%s_%s_slab_k%s_s%sv%s_%s" % (self.el, self.mpid, \
-                                                     self.k_product, self.ssize,
-                                                     self.vsize, self.reconstruction)
-
-    @property
-    def get_additional_fields(self):
-
-        # Get additional fields for raw data insertion that can
-        # not be derived from the vasp outputs like Miller index
-        # Return dict (additional fields)
-
-        additional_fields = {"structure_type": self.structure_type,
-                             "calculation_name": self.get_name,
-                             "conventional_spacegroup": \
-                                 {"symbol": self.sg.get_space_group_symbol(),
-                                  "number": self.sg.get_space_group_number()},
-                             "initial_structure": self.structure.as_dict(),
-                             "material_id": self.mpid}
-
-        if self.structure_type != "conventional_unit_cell":
-            additional_fields.update({"miller_index": self.hkl,
-                                      "scale_factor": self.scale_factor,
-                                       "reconstruction": self.reconstruction})
-
-        if self.structure_type == "slab_cell":
-            additional_fields.update({"oriented_unit_cell": self.ouc.as_dict(),
-                                      "slab_size":self.ssize, "shift": self.shift,
-                                      "vac_size": self.vsize})
-
-        return additional_fields
-
-    @property
-    def get_tasks(self):
-
-        # Get sequence of tasks depending on structure_type. This is similar
-        # to the OptimzeFW but has an additional task for generating children
-        # FWs as surface energy calculations require the transformation of a
-        # relaxed bulk structure into a slab
-        # Return list (tasks)
-
-        tasks = [CreateFolder(folder_name=os.path.join(self.cwd, self.get_name),
-                              change_dir=True, relative_path=True),
-                 WriteVaspFromIOSet(structure=self.structure,
-                                    vasp_input_set=self.get_input_set),
-                 RunVaspCustodian(vasp_cmd=self.vasp_cmd,
-                                  scratch_dir=self.scratch_dir,
-                                  auto_npar=">>auto_npar<<",
-                                  job_type="double_relaxation_run")]
-
-        if self.structure_type == "slab_cell":
-            tasks.append(RenameFile(file="LOCPOT.gz", new_name="LOCPOT.relax2.gz"))
-
-        tasks.append(VaspToDb(additional_fields=self.get_additional_fields,
-                              db_file=self.db_file))
-
-        if self.structure_type != "slab_cell":
-            tasks.append(FacetFWsGeneratorTask(structure_type=self.structure_type,
-                                               vasp_cmd=self.vasp_cmd, cwd=self.cwd,
-                                               db_file=self.db_file, mmi=self.mmi,
-                                               miller_index=self.hkl, mpid=self.mpid,
-                                               scratch_dir=self.scratch_dir,
-                                               k_product=self.k_product))
-
-        return tasks
-
-
-@explicit_serialize
-class FacetFWsGeneratorTask(FiretaskBase):
-    """
-    Task for generating FWs for oriented unit cell calculations or slab
-    calculations. If the initial structure is an oriented unit cell,
-    generate slab fws, if its a conventional unit cell, generate oriented
-    unit cell and reconstruction calculations.
-    """
-
-    required_params = ['structure_type', "scratch_dir", "k_product",
-                       "db_file", "vasp_cmd"]
-    optional_params = ["slab_gen_params", "mpid", "mmi", "miller_index", "cwd"]
-
-    def run_task(self, fw_spec):
-        """
-        Generates a list of oriented unit cell FWs or
-        slab cell FWs depending on the structure type.
-
-        Args:
-            structure_type (str): Type of structure. Options are:
-                conventional_unit_cell, oriented_unit_cell and slab_cell
-            scratch_dir (str): - if specified, uses this directory as the root
-                scratch dir. Supports env_chk.
-            k_product (int): Default to 50, kpoint number * length for a & b
-                directions, also for c direction in bulk calculations.
-            db_file (str): FULL path to file containing the database credentials.
-                Supports env_chk.
-            vasp_cmd (str): Command to run vasp.
-            slab_gen_params (dict): Parameters for SlabGenerator or generate_all_slabs
-            mpid (str): Materials Project ID of the conventional unit cell.
-            mmi (int): Max Miller index.
-            miller_index ([h, k, l]): Miller index of plane parallel to
-                surface (and oriented unit cell).
-            cwd (str): Location of directory to operate the
-                workflow and store the final outputs
-            \*\*kwargs: Other kwargs that are passed to Firework.__init__.
-        """
-
-        # get the parameters for SlabGenerator if we are getting a
-        # list of Slab calculation FWs or generate_all_slabs if we
-        # are getting a list of oritend unit cell calculation FWs
-        if self.get("slab_gen_params", None):
-            slab_gen_params = self.get("slab_gen_params", None)
-        else:
-            slab_gen_params = {"min_slab_size": 10, "min_vacuum_size": 10,
-                               "symmetrize": True, "center_slab": True}
-            slab_gen_params["max_normal_search"] = self.get("mmi") if \
-                self.get("mmi") else max(self.get("miller_index"))
-            if self.get("structure_type") == "conventional_unit_cell":
-                slab_gen_params["include_reconstructions"] = True
-
-        FWs = []
-        if self.get("structure_type") == "conventional_unit_cell":
-            # Then we create a set of FWs for oriented_unit_cell
-            # calculations and reconstructed slabs
-
-            slab_gen_params["structure"] = Structure.from_file("CONTCAR.relax2.gz")
-
-            all_slabs = generate_all_slabs(max_index=self.get("mmi"),
-                                           **slab_gen_params)
-            # Get all oriented unit cells up ot a max Miller index (mmi)
-            miller_list = []
-            for slab in all_slabs:
-                if slab.reconstruction:
-                    FWs.append(self.get_ouc_fw(slab))
-                else:
-                    # There are several surface terminations for an oriented unit
-                    # cell, we only need to calculate the oriented unit cell once
-                    if tuple(slab.miller_index) in miller_list:
-                        continue
-                    else:
-                        # build a oriented unit cell fw
-                        FWs.append(self.get_ouc_fw(slab))
-                        miller_list.append(tuple(slab.miller_index))
-
-        elif self.get("structure_type") == "oriented_unit_cell":
-
-            folder = os.path.basename(os.getcwd())
-            # If this is a reconstruction, we need to use the ReconstructionGenerator
-            if "_rec_" in folder:
-                ouc = Structure.from_file("CONTCAR.relax2.gz")
-                # ReconstructionGenerator only works on the conventional ucell
-                ucell = SpacegroupAnalyzer(ouc).get_conventional_standard_structure()
-                # Get the name of the reocnstruction
-                ns, n = "", 0
-                for s in folder:
-                    ns+=s
-                    if s == "_":
-                        n+=1
-                    if n == 5:
-                        break
-                rec = ReconstructionGenerator(ucell, slab_gen_params["min_slab_size"],
-                                              slab_gen_params["min_vacuum_size"],
-                                              reconstruction_name=folder[len(ns):])
-                FWs.append(self.get_slab_fw(rec.build_slab(), slab_gen_params))
-
-            else:
-                # Get the list of FWs for the various terminations
-                # for a slab based on the oriented unit cell
-                slab_gen_params["initial_structure"] = \
-                    Structure.from_file("CONTCAR.relax2.gz")
-                slab_gen_params["miller_index"] = [0,0,1]
-                symmetrize = slab_gen_params["symmetrize"]
-                del slab_gen_params["symmetrize"]
-                slabgen = SlabGenerator(**slab_gen_params)
-                for slab in slabgen.get_slabs(symmetrize=symmetrize):
-                    slab.miller_index = self.get("miller_index")
-                    FWs.append(self.get_slab_fw(slab, slab_gen_params))
-
-        return FWAction(additions=FWs)
-
-    def get_ouc_fw(self, slab):
-        """
-        Return a SurfCalcOptimizer FW for an oriented unit cell.
-
-        Args:
-            slab (Slab): Slab object containing various
-                attributes related to the slab
-        """
-
-        return SurfCalcOptimizer(slab.oriented_unit_cell, self.get("scratch_dir"),
-                                 self.get("k_product"), self.get("db_file"),
-                                 self.get("vasp_cmd"), "oriented_unit_cell",
-                                 reconstruction=slab.reconstruction,
-                                 miller_index=slab.miller_index,
-                                 scale_factor=slab.scale_factor,
-                                 mpid=self.get("mpid", "--"))
-
-    def get_slab_fw(self, slab, slab_gen_params):
-        """
-        Return a SurfCalcOptimizer FW for a Slab cell.
-
-        Args:
-            slab (Slab): Slab object containing various
-                attributes related to the slab
-            slab_gen_params (dict): Parameters for SlabGenerator
-                or generate_all_slabs
-        """
-
-        return SurfCalcOptimizer(slab, self.get("scratch_dir"),
-                                 self.get("k_product"), self.get("db_file"),
-                                 self.get("vasp_cmd"), "slab_cell",
-                                 miller_index=slab.miller_index,
-                                 scale_factor=slab.scale_factor,
-                                 ouc=slab.oriented_unit_cell, shift=slab.shift,
-                                 ssize=slab_gen_params["min_slab_size"],
-                                 vsize=slab_gen_params["min_vacuum_size"],
-                                 reconstruction=slab.reconstruction,
-                                 mpid=self.get("mpid", "--"))
